+++ example_scripts/c3_co2_response_v2.R
@@ -25,10 +25,10 @@ MAKE_VALIDATION_PLOTS <- TRUE
 MAKE_ANALYSIS_PLOTS <- TRUE
 
 # Decide whether to only keep points where stability conditions were met
-REQUIRE_STABILITY <- TRUE
+REQUIRE_STABILITY <- FALSE
 
 # Decide whether to remove some specific points
-REMOVE_SPECIFIC_POINTS <- TRUE
+REMOVE_SPECIFIC_POINTS <- FALSE
 
 # Choose a maximum value of Ci to use when fitting (ppm). Set to Inf to disable.
 MAX_CI <- Inf
@@ -48,7 +48,7 @@ PERFORM_STATS_TESTS <- TRUE
 # (Inf), then Cc = Ci and the resulting Vcmax values will be "apparent Vcmax,"
 # which is not solely a property of Rubisco and which may differ between plants
 # that have identical Vcmax but different gm.
-USE_GM_TABLE <- FALSE
+USE_GM_TABLE <- TRUE
 GM_VALUE <- Inf
 GM_UNITS <- "mol m^(-2) s^(-1) bar^(-1)"
 GM_TABLE <- list(
@@ -167,6 +167,19 @@ if (MAKE_VALIDATION_PLOTS) {
       ylab = paste('Net CO2 assimilation rate [', licor_data$units$A, ']')
     ))
 
+    # Plot all gsw-Ci curves in the data set
+    dev.new()
+    print(xyplot(
+      gsw ~ Ci | curve_identifier,
+      data = licor_data$main_data,
+      type = 'b',
+      pch = 16,
+      auto = TRUE,
+      grid = TRUE,
+      xlab = paste('Intercellular CO2 concentration [', licor_data$units$Ci, ']'),
+      ylab = paste('Stomatal conductance to H2O [', licor_data$units$gsw, ']')
+    ))
+
     # Make a plot to check humidity control
     dev.new()
     print(xyplot(
@@ -260,6 +273,9 @@ licor_data <- if (USE_GM_TABLE) {
 # Calculate total pressure (required for apply_gm)
 licor_data <- calculate_total_pressure(licor_data)
 
+# Calculate additional gas properties (required for calculate_c3_limitations)
+licor_data <- calculate_gas_properties(licor_data)
+
 # Calculate Cc
 licor_data <- apply_gm(licor_data)
 
@@ -281,6 +297,10 @@ c3_aci_results <- consolidate(by(
   cj_crossover_max = 800,                       # Wj must be < Wc when Cc > this value (ppm)
   fixed = c(NA, NA, NA, NA)
 ))
+
+# Calculate the relative limitations to assimilation (due to stomatal
+# conductance, mesophyll conductance, and biochemistry)
+c3_aci_results$fits <- calculate_c3_limitations(c3_aci_results$fits)
 
 if (MAKE_ANALYSIS_PLOTS) {
     # Plot the C3 A-Ci fits (including limiting rates)
@@ -353,9 +373,9 @@ if (REMOVE_STATISTICAL_OUTLIERS) {
 # Create a few data frames that will be helpful for plots and stats tests, and
 # make sure they are "factorized"
 
-all_samples_one_point <- licor_data[licor_data[, 'seq_num'] == POINT_FOR_BOX_PLOTS]
+all_samples_one_point <- c3_aci_results$fits[c3_aci_results$fits[, 'seq_num'] == POINT_FOR_BOX_PLOTS]
 aci_parameters <- c3_aci_results$parameters$main_data
-all_samples <- licor_data$main_data
+all_samples <- c3_aci_results$fits$main_data
 
 all_samples_one_point <- factorize_id_column(all_samples_one_point, EVENT_COLUMN_NAME)
 aci_parameters <- factorize_id_column(aci_parameters, EVENT_COLUMN_NAME)
@@ -379,12 +399,15 @@ if (MAKE_ANALYSIS_PLOTS) {
     xl <- "Genotype"
 
     plot_param <- list(
-      list(Y = all_samples_one_point[, 'A'],    X = x_s, xlab = xl, ylab = "Net CO2 assimilation rate (micromol / m^2 / s)",          ylim = c(0, 40),  main = boxplot_caption),
-      list(Y = all_samples_one_point[, 'iWUE'], X = x_s, xlab = xl, ylab = "Intrinsic water use efficiency (micromol CO2 / mol H2O)", ylim = c(0, 100), main = boxplot_caption),
-      list(Y = aci_parameters[, 'Vcmax_at_25'], X = x_v, xlab = xl, ylab = "Vcmax at 25 degrees C (micromol / m^2 / s)",              ylim = c(0, 200), main = fitting_caption),
-      list(Y = aci_parameters[, 'Rd_at_25'],    X = x_v, xlab = xl, ylab = "Rd at 25 degrees C (micromol / m^2 / s)",                 ylim = c(0, 3), main = fitting_caption),
-      list(Y = aci_parameters[, 'J_at_25'],     X = x_v, xlab = xl, ylab = "J at 25 degrees C (micromol / m^2 / s)",                  ylim = c(0, 225), main = fitting_caption),
-      list(Y = aci_parameters[, 'TPU'],         X = x_v, xlab = xl, ylab = "TPU (micromol / m^2 / s)",                                ylim = c(0, 30),  main = fitting_caption)
+      list(Y = all_samples_one_point[, 'A'],          X = x_s, xlab = xl, ylab = "Net CO2 assimilation rate (micromol / m^2 / s)",            ylim = c(0, 40),  main = boxplot_caption),
+      list(Y = all_samples_one_point[, 'iWUE'],       X = x_s, xlab = xl, ylab = "Intrinsic water use efficiency (micromol CO2 / mol H2O)",   ylim = c(0, 100), main = boxplot_caption),
+      list(Y = all_samples_one_point[, 'ls_rubisco'], X = x_s, xlab = xl, ylab = "Relative A limitation due to stomata (dimensionless)",      ylim = c(0, 0.5), main = boxplot_caption),
+      list(Y = all_samples_one_point[, 'lm_rubisco'], X = x_s, xlab = xl, ylab = "Relative A limitation due to mesophyll (dimensionless)",    ylim = c(0, 0.5), main = boxplot_caption),
+      list(Y = all_samples_one_point[, 'lb_rubisco'], X = x_s, xlab = xl, ylab = "Relative A limitation due to biochemistry (dimensionless)", ylim = c(0, 0.5), main = boxplot_caption),
+      list(Y = aci_parameters[, 'Vcmax_at_25'],       X = x_v, xlab = xl, ylab = "Vcmax at 25 degrees C (micromol / m^2 / s)",                ylim = c(0, 200), main = fitting_caption),
+      list(Y = aci_parameters[, 'Rd_at_25'],          X = x_v, xlab = xl, ylab = "Rd at 25 degrees C (micromol / m^2 / s)",                   ylim = c(0, 3),   main = fitting_caption),
+      list(Y = aci_parameters[, 'J_at_25'],           X = x_v, xlab = xl, ylab = "J at 25 degrees C (micromol / m^2 / s)",                    ylim = c(0, 225), main = fitting_caption),
+      list(Y = aci_parameters[, 'TPU'],               X = x_v, xlab = xl, ylab = "TPU (micromol / m^2 / s)",                                  ylim = c(0, 30),  main = fitting_caption)
     )
 
     if (INCLUDE_FLUORESCENCE) {
@@ -410,22 +433,26 @@ if (MAKE_ANALYSIS_PLOTS) {
     rc_caption <- "Average response curves for each event"
 
     x_ci <- all_samples[, 'Ci']
+    x_cc <- all_samples[, 'Cc']
     x_s <- all_samples[, 'seq_num']
     x_e <- all_samples[, EVENT_COLUMN_NAME]
 
     ci_lim <- c(-50, 1500)
+    cc_lim <- c(-50, 1500)
     a_lim <- c(-10, 55)
-    gs_lim <- c(0, 0.7)
+    gsw_lim <- c(0, 0.7)
     phi_lim <- c(0, 0.4)
 
     ci_lab <- "Intercellular [CO2] (ppm)"
+    cc_lab <- "Mesophyll [CO2] (ppm)"
     a_lab <- "Net CO2 assimilation rate (micromol / m^2 / s)\n(error bars: standard error of the mean for same CO2 setpoint)"
-    gs_lab <- "Stomatal conductance to CO2 (mol / m^2 / s)\n(error bars: standard error of the mean for same CO2 setpoint)"
+    gsw_lab <- "Stomatal conductance to H2O (mol / m^2 / s)\n(error bars: standard error of the mean for same CO2 setpoint)"
     phi_lab <- "PhiPSII (dimensionless)\n(error bars: standard error of the mean for same CO2 setpoint)"
 
     avg_plot_param <- list(
-        list(all_samples[, 'A'],   x_ci, x_s, x_e, xlab = ci_lab, ylab = a_lab,  xlim = ci_lim, ylim = a_lim),
-        list(all_samples[, 'gsw'], x_ci, x_s, x_e, xlab = ci_lab, ylab = gs_lab, xlim = ci_lim, ylim = gs_lim)
+        list(all_samples[, 'A'],   x_ci, x_s, x_e, xlab = ci_lab, ylab = a_lab,   xlim = ci_lim, ylim = a_lim),
+        list(all_samples[, 'A'],   x_cc, x_s, x_e, xlab = cc_lab, ylab = a_lab,   xlim = cc_lim, ylim = a_lim),
+        list(all_samples[, 'gsw'], x_ci, x_s, x_e, xlab = ci_lab, ylab = gsw_lab, xlim = ci_lim, ylim = gsw_lim)
     )
 
     if (INCLUDE_FLUORESCENCE) {